requirements.txt
qrcode[pil]
pillow
segno
